Move repository concern from application layer to domain layer
