+++ lib/interfaces/storage/UserRepositoryInMemory.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const User = require('../../domain/User');
-const UserRepository = require('../../application/repositories/UserRepository');
+const UserRepository = require('../../domain/UserRepository');
 
 module.exports = class extends UserRepository {
