+++ lib/interfaces/storage/UserRepositoryMongo.js
@@ -2,7 +2,7 @@
 
 const User = require('../../domain/User');
 const MongooseUser = require('../../infrastructure/orm/mongoose/schemas/User');
-const UserRepository = require('../../application/repositories/UserRepository');
+const UserRepository = require('../../domain/UserRepository');
 
 module.exports = class extends UserRepository {
