+++ lib/interfaces/storage/UserRepositorySQLite.js
@@ -2,7 +2,7 @@
 
 const sequelize = require('../../infrastructure/orm/sequelize/sequelize');
 const User = require('../../domain/User');
-const UserRepository = require('../../application/repositories/UserRepository');
+const UserRepository = require('../../domain/UserRepository');
 
 module.exports = class extends UserRepository {
